Refactor InMemorySubscriptionStore
Refactored store out into seperate file from main.go

added function to return a new in memory store so a new instance of the struct is used each time
